test: move mocks to their own directory
