+++ tests/mocks/nextjs.ts
